# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Учёба с Лучиком API",
    "version": "0.1.0",
    "description": "Student mini-app API. MAX initData is exchanged for a short-lived bearer session. Schedules are synthetic, and lesson generation uses RouterAI."
  },
  "paths": {
...
    "/api/admin/login": {
      "post": {
        "summary": "Login",
        "operationId": "login_api_admin_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-Admin-Action",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "const": "1"
            }
          }
        ]
      }
    },
    "/api/admin/logout": {
      "post": {
        "summary": "Logout",
        "operationId": "logout_api_admin_logout_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "AdminSession": []
          }
        ],
        "parameters": [
          {
            "name": "X-Admin-Action",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "const": "1"
            }
          }
        ]
      }
    },
    "/api/admin/me": {
      "get": {
        "summary": "Me",
        "operationId": "me_api_admin_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "AdminSession": []
          }
        ]
      }
    },
    "/api/admin/students": {
      "get": {
        "summary": "Students",
        "operationId": "students_api_admin_students_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "AdminSession": []
          }
        ]
      }
    },
    "/api/admin/students/{uid}/schedule": {
      "get": {
        "summary": "Schedule",
        "operationId": "schedule_api_admin_students__uid__schedule_get",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Uid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "AdminSession": []
          }
        ]
      },
      "post": {
        "summary": "Add Entry",
        "operationId": "add_entry_api_admin_students__uid__schedule_post",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Uid"
            }
          },
          {
            "name": "X-Admin-Action",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "const": "1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleEntry"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "AdminSession": []
          }
        ]
      }
    },
    "/api/admin/students/{uid}/schedule/{entry_id}": {
      "put": {
        "summary": "Update Entry",
        "operationId": "update_entry_api_admin_students__uid__schedule__entry_id__put",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Uid"
            }
          },
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Entry Id"
            }
          },
          {
            "name": "X-Admin-Action",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "const": "1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleEntry"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "AdminSession": []
          }
        ]
      },
      "delete": {
        "summary": "Delete Entry",
        "operationId": "delete_entry_api_admin_students__uid__schedule__entry_id__delete",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Uid"
            }
          },
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Entry Id"
            }
          },
          {
            "name": "X-Admin-Action",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "const": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "AdminSession": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminLogin": {
        "properties": {
          "password": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "password"
        ],
        "title": "AdminLogin"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScheduleEntry": {
        "properties": {
          "day": {
            "type": "string",
            "format": "date",
            "title": "Day"
          },
          "period": {
            "type": "integer",
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "Period"
          },
          "subject": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Subject"
          },
          "topic": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Topic"
          },
          "homework": {
            "type": "string",
            "maxLength": 1000,
            "title": "Homework"
          },
          "assessment": {
            "type": "boolean",
            "title": "Assessment",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "day",
          "period",
          "subject",
          "topic",
          "homework"
        ],
        "title": "ScheduleEntry"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "StudentSession": {
        "type": "http",
        "scheme": "bearer",
        "description": "Session returned by POST /api/auth/max; valid for 24 hours."
      },
      "AdminSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "luchik_admin",
        "description": "Separate admin session; only needed for schedule administration."
      },
      "MaxWebhookSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Max-Bot-Api-Secret"
      }
    }
  },
  "servers": [
    {
      "url": "https://85-159-231-99.sslip.io",
      "description": "Public hackathon deployment"
    }
  ]
}